requirements.txt
# THE EYE - Dependencies
# ========================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Camera (Raspberry Pi)
# picamera2>=0.3.12  # Uncomment on Raspberry Pi

# Computer Vision
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# Machine Learning
# tensorflow>=2.15.0  # Heavy - install separately on Pi
# tflite-runtime>=2.14.0  # Lighter alternative for Pi

# Face Recognition
# face-recognition>=1.3.0  # Optional - requires dlib

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Database
aiosqlite>=0.19.0

# Home Automation
# homeassistant-api>=4.0.0  # Optional - for HA integration
# paho-mqtt>=1.6.1  # Optional - for MQTT

# GPIO (Raspberry Pi only)
# RPi.GPIO>=0.7.1  # Uncomment on Raspberry Pi
# gpiozero>=2.0  # Uncomment on Raspberry Pi

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# UI (Optional)
streamlit>=1.31.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
